accept Mode from Phone

+++ CoolFontWin/Program.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Runtime.InteropServices;
+using System.IO;
 using System.Diagnostics;
 using System.Collections.Generic;
 using System.Linq;
@@ -13,8 +15,28 @@ namespace CoolFontWin
 {
     class Program
     {
+        [DllImport("Kernel32")]
+        public static extern bool SetConsoleCtrlHandler(HandlerRoutine Handler, bool Add);
+
+        // A delegate type to be used as the handler routine 
+        // for SetConsoleCtrlHandler.
+        public delegate bool HandlerRoutine(CtrlTypes CtrlType);
+
+        // An enumerated type for the control messages
+        // sent to the handler routine.
+        public enum CtrlTypes
+        {
+            CTRL_C_EVENT = 0,
+            CTRL_BREAK_EVENT,
+            CTRL_CLOSE_EVENT,
+            CTRL_LOGOFF_EVENT = 5,
+            CTRL_SHUTDOWN_EVENT
+        }
+
+
         static void Main(string[] args)
-        {           
+        {
+            SetConsoleCtrlHandler(new HandlerRoutine(ConsoleCtrlCheck), true);
             int tryport = ProcessArgs(args);
 
             if (tryport == 0)
@@ -48,41 +70,37 @@ namespace CoolFontWin
             CoolFontSimulator sim = new CoolFontSimulator(Config.Mode); // will change Mode if necessary
 
             string rcvd;
-            char[] delimiters = { ':' };
             int[] vals = { 0, 0, 0, 0 };
             int[] vals_last = vals;
             int buttons = 0; // bitmask
+            int modeIn = (int)Config.Mode;
             int T = 0; // total time
             int timeout = 30; // set to -1 to block on every socket read
             int tries = timeout + 1;
+
             sim.logOutput = false;
+            bool logRcvd = false;
 
 
             //TODO: execute loop in background thread and allow user to break out
             while (true)
             {
-                if (IsKeyPressed(ConsoleKey.Escape))
-                {
-                    break;
-                }
-
-                // Receive data from iPhone, parse it, and translate it to the correct inputs
-                /* vals[0]: 0-1000: represents user running at 0 to 100% speed.
-                 * vals[1]: 0-360,000: represents the direction user is facing (in degrees)
-                 * vals[2]: always 0
-                 * vals[3]: -infinity to infinity: user rotation rate in radians per second (x1000)
-                 */
-
                 // do not block, returns "" if nothing to rcv
                 //rcvd = listener.receiveStringAsync();
                 rcvd = listener.pollSocket(Config.socketPollInterval);
+
+
                 try
                 {
-                    vals = listener.parseString2Ints(rcvd, delimiters);
-                    buttons = listener.parseButtons(rcvd, delimiters);
+                    vals = listener.parseString2Ints(rcvd);
+                    buttons = listener.parseButtons(rcvd);
+                    modeIn = listener.parseMode(rcvd, (int)Config.Mode); // Config.Mode is a fallback
                     vals = Algorithm.LowPassFilter(vals,vals_last,Config.RCFilterStrength,Config.dt);
                     vals_last = vals;
                     tries = 0;
+
+                    if (logRcvd && (T % 20 == 0))
+                        Console.WriteLine("{0} \n {1}{2}{3} \n {4}",modeIn, vals[0], vals[1], vals[2] , buttons);
                 }
                 catch
                 {
@@ -95,31 +113,36 @@ namespace CoolFontWin
                 }
 
                 /* Get input from connected XInput device */
-                sim.AddValues(vals, Config.Mode);
-               // sim.AddButtons(buttons, Config.Mode);
+
+
+                sim.UpdateMode(modeIn);
+                sim.AddValues(vals);
+                sim.AddButtons(buttons);
 
                 if (dev.controller != null && dev.controller.IsConnected)
                 {
                     state = dev.controller.GetState();
                     sim.AddControllerState(state);
                 }
-                sim.AddButtons(buttons, Config.Mode);
+
                 sim.FeedVJoy();
                 sim.ResetValues();
-                if (sim.logOutput) { Console.Write(" ({0}) \n", tries); }
+                if (sim.logOutput)
+                {
+                    Console.Write(" ({0}) \n", tries);
+                }
                 T++;
-            }
 
-            sim.DisableVJoy(Config.ID);
-            JavaProc.Kill();
-            listener.Close();
-            dev.controller = null;
-            dev = null;
+            }
         }
 
-        public static bool IsKeyPressed(ConsoleKey key)
+        private static bool ConsoleCtrlCheck(CtrlTypes ctrlType)
         {
-            return false; // Console.KeyAvailable && Console.ReadKey(true).Key == key;
+            if (ctrlType == CtrlTypes.CTRL_CLOSE_EVENT)
+            {
+                JavaProc.Kill();
+            }
+            return true;
         }
 
         static int ProcessArgs(string[] args)
